requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.0
requests==1.1.0
